# src/internal/database/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package database

import (
	"database/sql"
	"time"
)

type MediaContent struct {
	ID          int32
	Title       string
	Description sql.NullString
	Body        string
	ContentType string
	CreatedAt   time.Time
	UserID      int32
}

type User struct {
	ID        int32
	FirstName string
	LastName  string
	Email     string
	Password  string
}